refactor(a1): extract list/read routes to routers/lists.py (BR-CODE-1 slice 3)
Moves /api/watchlist (GET), /api/runs, /api/snapshots/latest, and
/api/snapshots/velocity out of app.py into routers/lists.py. Factory takes a
live require_sb getter + require_auth, so handlers resolve the current
require_sb at request time — auth/db ownership stays in app.py and the
monkeypatch tests (test_watchlist_runs_snapshots.py) keep working. Watchlist
POST/DELETE (mutating) stay in app.py. 37 tests green; entrypoint imports
clean; moved routes resolve a post-build monkeypatched require_sb.
