+++ KANBAN.md
@@ -48,7 +48,7 @@ Prioritized program from the 2026-06-19 cross-axis bottleneck audit (code · dat
 |---|---|---|---|---|
 | **P0** | → `A1-OPS-3` | A1 / Op | **SG broker token is the #1 product ceiling** — one token shared with an invisible external prod program; 63–77% 429, 1,235 starved events, 10/30 MLB games 2–3d stale. Physics limit, not code. | See `A1-OPS-3` (OPS table). **Operator action:** coordinate SG token isolation from the external app; interim approve the `p_max 8→4` cron-body cut. |
 | **P0** | → `A1-OPS-16..21` | A1 / Op | **Snapshot firehose storage + bloat** — `listings_snapshots` 3.6GB/8.96M, SG listings 3.2GB, SG sales 1.18GB (11–23× overcount), `espn_injuries` 175MB/16.5k rows. Retention Waves 2–4 (dedup · change-detection write-gate · partition+VACUUM) still pending. | See `A1-OPS-16/17/18/21`. Finish Wave 3 change-detection (in progress), then Wave 4 partitioning. |
-| **P1** | **BR-CODE-1** | D0 / A1 / Op [IN PROGRESS] | **`app.py` monolith — 8,753 LOC, 112 routes, 236 fns, 302 inline SQL calls, ALL surfaces in one file** (5 fns >220 lines). #1 velocity drag; every route re-writes the same filter chains, no DB-access layer. | Carve out a `routers/` package by domain (broker · store · seatgeek · seatdata · axs · admin) + a thin shared `db`/query helper; move routes incrementally behind unchanged paths. Land in small PRs, tests green each step. **Slice 1 (done):** `routers/` package + `routers/site_essentials.py` (robots/sitemap/favicon, factory takes deploy base URL). **Slice 2 (done):** `routers/catalog.py` — TEvo performer/venue/configuration detail passthroughs, factory takes a **live `client` getter + `require_auth`** (auth code stays in app.py untouched; monkeypatch tests preserved; covered by `test_catalog_routes.py`). **Next slices:** `/api/public/config`+`/version.json`+`/healthz` → `routers/public_meta.py`; then a shared `core` (auth+db+client) before the big broker/store blocks. |
+| **P1** | **BR-CODE-1** | D0 / A1 / Op [IN PROGRESS] | **`app.py` monolith — 8,753 LOC, 112 routes, 236 fns, 302 inline SQL calls, ALL surfaces in one file** (5 fns >220 lines). #1 velocity drag; every route re-writes the same filter chains, no DB-access layer. | Carve out a `routers/` package by domain (broker · store · seatgeek · seatdata · axs · admin) + a thin shared `db`/query helper; move routes incrementally behind unchanged paths. Land in small PRs, tests green each step. **Slice 1 (done):** `routers/` package + `routers/site_essentials.py` (robots/sitemap/favicon, factory takes deploy base URL). **Slice 2 (done):** `routers/catalog.py` — TEvo performer/venue/configuration detail passthroughs (live `client` getter + `require_auth`). **Slice 3 (done):** `routers/lists.py` — `/api/watchlist` GET + `/api/runs` + `/api/snapshots/latest|velocity` (live `require_sb` getter + `require_auth`; watchlist POST/DELETE stay in app.py; covered by `test_watchlist_runs_snapshots.py`). **Next slices:** `/api/public/config`+`/version.json`+`/healthz` → `routers/public_meta.py`; then a shared `core` (auth+db+client) before the big broker/store blocks. |
 | **P1** | **BR-CODE-2** | A1 / D0 | **9 `*_client.py` copy-paste ~100 LOC each** (read-only guard, retry/backoff, vault-secret, session/headers) with *inconsistent* retry → uneven rate-limit handling (ties into the SG 429 story). | Extract a `BaseReadOnlyClient` (GET-only guard + unified exponential backoff honoring `ratelimit-remaining` + vault resolver); migrate clients one at a time. ~900 LOC reclaimed; `tests/test_readonly_guards.py` stays green. |
 | **P2** | **BR-CODE-3** | B1 / D0 [IN PROGRESS, PR #584] | **Test coverage ~12%** (5,951 test LOC vs ~46.5k); **60+ broker/admin routes untested** (`/api/broker/*/chart-data`, `/movers`, all `/api/admin/*`, `/api/seatdata/*`, `/api/axs/*`). | Add route-level tests for the broker/admin surface (mock the SECDEF RPCs); prioritize the event-page + movers paths. Target 30% ratio. **Increment 1 (done):** `tests/test_public_infra_routes.py` — 10 tests over the public non-DB infra routes. **Increment 2 (done):** `tests/test_broker_routes.py` — 5 tests over `/api/broker/leagues`, `/performer/{id}/assets`, `/event/{id}/overview` + a reusable `FakeSupabase`. **Increment 3 (done):** `tests/test_broker_movers.py` — 4 tests over `/api/broker/movers` v2 + `/watchlist-movers`. **Increment 4 (done):** `tests/test_watchlist_runs_snapshots.py` — 6 tests over `/api/watchlist`, `/api/runs`, `/api/snapshots/latest` (event-join enrichment + sort), `/api/snapshots/velocity`. **Increment 5 (done):** `tests/test_catalog_routes.py` — 6 tests over the TEvo-catalog passthrough GETs (`/api/performers/{id}`, `/api/venues/{id}`, `/api/configurations` + `/{id}`) asserting arg-passthrough + response relay. **Total: 31 route tests.** **Next:** `/event/{id}/chart-data` (469-line fn) + the `/api/admin/*` POST orchestrators (mock the upstream fire). |
 | **P2** | → `A1-OPS-22` | A1 / Op | **AQ map duplicates/false-matches** — 235 `tevo_event_id`s with ≥2 rows → cross-source visibility gaps; safe-class consolidation fn built, pending apply. | See `A1-OPS-22`. Operator apply + dry-run review → execute. |
